Create Recode Data Base

Create and store a recode data base in an Excel file. The Excel file will have sheets corresponding to each recodeList name.

Usage

createRecodeDB(recodeListList, filePath)

Arguments

recodeListList

A named list of recodeLists.

filePath

Path to the .xlsx file in which the data base should be stored.

Examples

# create a named list of data frames
recodeListList <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                        Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
recodeListList
#> $Europe
#>   id  oldValues newValues
#> 1  1     Berlin   Germany
#> 2  2 Copenhagen   Denmark
#> 3  3       Rome     Italy
#> 4  4     Madrid     Spain
#> 
#> $Asia
#>   id oldValues  newValues
#> 1  1      Baku Azerbaijan
#> 2  2     Tokyo      Japan
#> 3  3 Kathmandu      Nepal
#> 4  4 Singapore  Singapore
#> 
filePath_temp <- tempfile(fileext = ".xlsx")
# create the Excel-file
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp)
#> NULL